+++ server/controller/imageController.js
@@ -1,6 +1,7 @@
-const Albums = require("../model/albumsSchema");
-const Images = require("../model/imagesSchema");
-const AppError = require("../utils/appError");
+const { validationResult } = require('express-validator');
+const Albums = require('../model/albumsSchema');
+const Images = require('../model/imagesSchema');
+const AppError = require('../utils/appError');
 
 exports.getImagesByAlbumId = async (req, res, next) => {
 	const albumId = req.params.id;
@@ -10,23 +11,23 @@ exports.getImagesByAlbumId = async (req, res, next) => {
 		existingAlbum = await Albums.findById(albumId);
 	} catch (error) {
 		return next(
-			new AppError("Fetching album failed, please try again later.", 404)
+			new AppError('Fetching album failed, please try again later.', 404)
 		);
 	}
 
 	if (!existingAlbum) {
-		return next(new AppError("Could not find album for provided id", 404));
+		return next(new AppError('Could not find album for provided id', 404));
 	}
 
 	let albumImages;
 	try {
 		albumImages = await Images.find({ albumId: albumId });
 	} catch (error) {
-		return new AppError("Fetching user failed, please try again later.", 404);
+		return new AppError('Fetching user failed, please try again later.', 404);
 	}
 
 	if (!albumImages || albumImages.length === 0) {
-		return new AppError("Could not find images for the this album", 404);
+		return new AppError('Could not find images for the this album', 404);
 	}
 
 	res.status(200).json({
@@ -37,19 +38,17 @@ exports.getImagesByAlbumId = async (req, res, next) => {
 exports.getImageById = async (req, res, next) => {
 	const imageId = req.params.id;
 
-	console.log(imageId);
-
 	let image;
 	try {
 		image = await Images.findById(imageId);
 	} catch (error) {
 		return next(
-			new AppError("Fetching image failed, please try again later", 401)
+			new AppError('Fetching image failed, please try again later', 401)
 		);
 	}
 
 	if (!image) {
-		return next(new AppError("Could not find image ", 404));
+		return next(new AppError('Could not find image ', 404));
 	}
 
 	res.status(200).json({
@@ -60,16 +59,28 @@ exports.getImageById = async (req, res, next) => {
 exports.createImages = async (req, res, next) => {
 	let album;
 
+	const errors = validationResult(req);
+
+	if (!errors.isEmpty()) {
+		return next(
+			new AppError('Invalid inputs passed, please check your data.', 400)
+		);
+	}
+
 	try {
 		album = await Albums.findById(req.body.albumId);
 	} catch (error) {
 		return next(
-			new AppError("Fetching album failed, please try again later", 500)
+			new AppError('Fetching album failed, please try again later', 500)
 		);
 	}
 
 	if (!album) {
-		return next(new AppError("Could not find album for provided id", 404));
+		return next(new AppError('Could not find album for provided id', 404));
+	}
+
+	if (req.user._id.toString() !== album.creator.toString()) {
+		return next(new AppError('You dont have permission to create image', 401));
 	}
 
 	try {
@@ -83,7 +94,7 @@ exports.createImages = async (req, res, next) => {
 		});
 	} catch (error) {
 		return next(
-			new AppError("Could not create image, please check your credentials", 404)
+			new AppError('Could not create image, please check your credentials', 404)
 		);
 	}
 };
@@ -96,12 +107,18 @@ exports.deleteAllImages = async (req, res, next) => {
 		album = await Albums.findById(albumId);
 	} catch (error) {
 		return next(
-			new AppError("Fetching album failed, please try again later", 500)
+			new AppError('Fetching album failed, please try again later', 500)
 		);
 	}
 
 	if (!album) {
-		return next(new AppError("Could not find album for provided id", 404));
+		return next(new AppError('Could not find album for provided id', 404));
+	}
+
+	if (req.user._id.toString() !== album.creator.toString()) {
+		return next(
+			new AppError('You dont have permission to delete this album', 401)
+		);
 	}
 
 	try {
@@ -112,22 +129,37 @@ exports.deleteAllImages = async (req, res, next) => {
 		});
 	} catch (error) {
 		return next(
-			new AppError("Something went wrong, Could not delete images", 404)
+			new AppError('Something went wrong, Could not delete images', 404)
 		);
 	}
 };
 exports.deleteImageById = async (req, res, next) => {
 	const imageId = req.params.id;
 
+	let album;
+	try {
+		const image = await Images.findById(imageId);
+		album = image.albumId;
+	} catch (error) {
+		return next(
+			new AppError('Something went wrong, Could not find images', 404)
+		);
+	}
+
+	if (req.user._id.toString() !== album.creator.toString()) {
+		return next(
+			new AppError('You dont have permission to delete this image', 401)
+		);
+	}
+
 	try {
 		await Images.deleteImageById(imageId);
-
 		res.status(200).json({
 			image: null,
 		});
 	} catch (error) {
 		return next(
-			new AppError("Something went wrong, Could not find images", 404)
+			new AppError('Something went wrong, Could not find images', 404)
 		);
 	}
 };
